ledger/sync.py
import os
import plaid
from plaid.api import plaid_api
from plaid.model.transactions_get_request import TransactionsGetRequest
from plaid.model.transactions_get_request_options import TransactionsGetRequestOptions
import datetime
from ledger.database import save_transaction

def sync_account():
    # Placeholder for Plaid Sync Logic
    print("Connecting to Plaid Development Environment...")
    # Real implementation requires Access Token generation via Link
